测试修改了访客接口api

+++ src/main/java/com/peashoot/blog/controller/ArticleController.java
@@ -3,6 +3,7 @@ package com.peashoot.blog.controller;
 import com.peashoot.blog.aspect.annotation.ErrorRecord;
 import com.peashoot.blog.aspect.annotation.VisitLimit;
 import com.peashoot.blog.batis.entity.ArticleDO;
+import com.peashoot.blog.batis.entity.RoleDO;
 import com.peashoot.blog.batis.entity.SysUserDO;
 import com.peashoot.blog.batis.enums.VisitActionEnum;
 import com.peashoot.blog.batis.entity.OperateRecordDO;
@@ -15,6 +16,7 @@ import com.peashoot.blog.context.request.article.ChangedArticleDTO;
 import com.peashoot.blog.context.response.ApiResp;
 import com.peashoot.blog.context.response.article.ArticleIntroductionDTO;
 import com.peashoot.blog.context.response.article.ArticlesCollectionDTO;
+import com.peashoot.blog.util.IpUtils;
 import com.peashoot.blog.util.SecurityUtil;
 import com.peashoot.blog.util.StringUtils;
 import io.swagger.annotations.Api;
@@ -22,9 +24,12 @@ import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.util.Date;
 import java.util.List;
@@ -45,18 +50,21 @@ public class ArticleController {
     /**
      * Article操作类
      */
-    @Autowired
-    private ArticleService articleService;
+    private final ArticleService articleService;
     /**
      * 系统用户信息操作类
      */
-    @Autowired
-    private SysUserService sysUserService;
+    private final SysUserService sysUserService;
     /**
      * 访客操作记录操作类
      */
-    @Autowired
-    private OperateRecordService visitRecordService;
+    private final OperateRecordService visitRecordService;
+
+    public ArticleController(ArticleService articleService, SysUserService sysUserService, OperateRecordService visitRecordService) {
+        this.articleService = articleService;
+        this.sysUserService = sysUserService;
+        this.visitRecordService = visitRecordService;
+    }
 
     /**
      * 根据查询条件获取符合条件的Article
@@ -66,47 +74,59 @@ public class ArticleController {
      */
     @RequestMapping(path = "list")
     @ApiOperation("获取符合查询条件的文章")
-    public ApiResp<ArticlesCollectionDTO> getArticles(@RequestBody @Valid ArticleSearchDTO apiReq, BindingResult bindingResult) {
+    public ApiResp<ArticlesCollectionDTO> getArticles(@RequestBody @Validated ArticleSearchDTO apiReq) {
         List<ArticleDO> matchedList = articleService.listPagedArticles(apiReq.getPageSize(), apiReq.getPageIndex(), apiReq.getAuthorLike(), apiReq.getKeywordLike(), apiReq.getTitleLike());
         int totalCount = articleService.countTotalRecords(apiReq.getAuthorLike(), apiReq.getKeywordLike(), apiReq.getTitleLike());
         ApiResp<ArticlesCollectionDTO> retResp = new ApiResp<ArticlesCollectionDTO>().success();
         ArticlesCollectionDTO data = new ArticlesCollectionDTO();
-        data.setArticleList(matchedList.stream().map(a -> ArticleIntroductionDTO.createArticlesInfo(a)).collect(Collectors.toList()));
+        data.setArticleList(matchedList.stream().map(ArticleIntroductionDTO::createArticlesInfo).collect(Collectors.toList()));
         data.setPageSize(apiReq.getPageSize());
         data.setPageIndex(apiReq.getPageIndex());
         data.setTotalRecordsCount(totalCount);
+        retResp.setData(data);
         return retResp;
     }
 
     /**
-     * 新增或更新Article
+     * 新增或更新Article(只有文章作者和文管员可以更新)
      *
      * @param apiReq 需要变更的Article
      * @return 是否成功
      */
     @PostMapping(path = "modify")
     @ApiOperation("新增或更新文章")
-    @PreAuthorize("hasRole('writer')")
-    public ApiResp<Boolean> insertOrUpdateArticle(@RequestBody ChangedArticleDTO apiReq) {
+    @PreAuthorize("hasAuthority('article_modify')")
+    public ApiResp<Boolean> insertOrUpdateArticle(@RequestBody @Validated ChangedArticleDTO apiReq) {
         ApiResp<Boolean> retResp = new ApiResp<>();
-        retResp.setCode(501);
+        retResp.setCode(ApiResp.PROCESS_ERROR);
         retResp.setMessage("Failure to save article.");
         boolean isInsert = StringUtils.isNullOrEmpty(apiReq.getId());
         ArticleDO articleEntity = isInsert ? new ArticleDO() : articleService.selectById(apiReq.getId());
-        apiReq.copyTo(articleEntity);
+        if(articleEntity == null) {
+            retResp.setCode(ApiResp.NO_RECORD_MATCH);
+            retResp.setMessage("No match record");
+            return retResp;
+        }
         SysUserDO curUser = SecurityUtil.getCurrentUser();
         if (curUser == null) {
             return retResp;
         }
+        apiReq.copyTo(articleEntity);
         articleEntity.setModifyUserId(sysUserService.getIdByUsername(curUser.getUsername()));
+        articleEntity.setModifyTime(new Date());
         boolean result;
         if (isInsert) {
             articleEntity.setCreateTime(articleEntity.getModifyTime());
             articleEntity.setCreateUserId(articleEntity.getModifyUserId());
-            visitRecordService.insertNewRecordAsync(articleEntity.getCreateUserId(), apiReq.getId(), VisitActionEnum.CREATE_ARTICLE, new Date(), "Add an article：" + apiReq.getTitle());
+            visitRecordService.insertNewRecordAsync(articleEntity.getCreateUserId(), apiReq.getId(), apiReq.getVisitorIP(), VisitActionEnum.CREATE_ARTICLE, new Date(), "Add an article：" + apiReq.getTitle());
             result = articleService.insert(articleEntity) > 0;
+        } else if (!curUser.getUsername().equals(articleEntity.getCreateUser().getUsername())
+                && curUser.getAuthorities().stream().noneMatch(p -> RoleDO.ROLE_OF_ARTICLE_MANAGER.equalsIgnoreCase(p.getAuthority()))) {
+            retResp.setCode(ApiResp.BAD_REQUEST);
+            retResp.setMessage("No permission to change");
+            return retResp;
         } else {
-            visitRecordService.insertNewRecordAsync(articleEntity.getModifyUserId(), apiReq.getId(), VisitActionEnum.UPDATE_ARTICLE, new Date(), "Modify an article：" + apiReq.getTitle());
+            visitRecordService.insertNewRecordAsync(articleEntity.getModifyUserId(), apiReq.getId(), apiReq.getVisitorIP(), VisitActionEnum.UPDATE_ARTICLE, new Date(), "Modify an article：" + apiReq.getTitle());
             result = articleService.update(articleEntity) > 0;
         }
         if (result) {
@@ -123,16 +143,28 @@ public class ArticleController {
      */
     @PostMapping(path = "delete")
     @ApiOperation("删除文章")
-    @PreAuthorize("hasRole('writer')")
-    public ApiResp<Boolean> deleteArticle(@RequestParam("articleId") String id) {
+    @PreAuthorize("hasAuthority('article_remove')")
+    public ApiResp<Boolean> deleteArticle(HttpServletRequest request, @RequestParam("articleId") String id) {
         ApiResp<Boolean> retResp = new ApiResp<>();
-        retResp.setCode(501);
+        retResp.setCode(ApiResp.PROCESS_ERROR);
         retResp.setMessage("Failure to remove article.");
+        ArticleDO articleEntity = articleService.selectById(id);
+        if(articleEntity == null) {
+            retResp.setCode(ApiResp.NO_RECORD_MATCH);
+            retResp.setMessage("No match record");
+            return retResp;
+        }
         SysUserDO curUser = SecurityUtil.getCurrentUser();
         if (curUser == null) {
             return retResp;
         }
-        visitRecordService.insertNewRecordAsync(sysUserService.getIdByUsername(curUser.getUsername()), id, VisitActionEnum.DELETE_ARTICLE, new Date(), "Delete an article, it's id is " + id);
+        if (!curUser.getUsername().equals(articleEntity.getCreateUser().getUsername())
+                && curUser.getAuthorities().stream().noneMatch(p -> RoleDO.ROLE_OF_ARTICLE_MANAGER.equalsIgnoreCase(p.getAuthority()))) {
+            retResp.setCode(ApiResp.BAD_REQUEST);
+            retResp.setMessage("No permission to change");
+            return retResp;
+        }
+        visitRecordService.insertNewRecordAsync(sysUserService.getIdByUsername(curUser.getUsername()), id, IpUtils.getIpAddr(request), VisitActionEnum.DELETE_ARTICLE, new Date(), "Delete an article, it's id is " + id);
         // 逻辑删除设置IsDelete为true
         boolean result = articleService.remove(id) > 0;
         if (result) {
@@ -152,7 +184,7 @@ public class ArticleController {
     @VisitLimit(value = 5)
     public ApiResp<Boolean> agreeOrDisagreeArticle(@RequestBody ArticleAgreeDTO apiReq) {
         ApiResp<Boolean> resp = new ApiResp<>();
-        resp.setCode(406);
+        resp.setCode(ApiResp.PROCESS_ERROR);
         resp.setMessage("Failure to action or disagree article");
         OperateRecordDO visitRecordDO = visitRecordService.selectLastRecordByVisitorIdAndArticleId(apiReq.getVisitorId(), apiReq.getArticleId());
         int agree = 0, disagree = 0;
@@ -190,9 +222,10 @@ public class ArticleController {
             return resp;
         }
         // 新增访客操作记录并修改评论点赞反对数
-        visitRecordService.insertNewRecordAsync(apiReq.getVisitorId(), apiReq.getArticleId(), apiReq.getAction(), new Date(), "");
-        boolean result = articleService.updateSupportAndDisagreeState(apiReq.getArticleId(), agree, disagree);
-        resp.success().setData(result);
+        visitRecordService.insertNewRecordAsync(apiReq.getVisitorId(), apiReq.getArticleId(), apiReq.getVisitorIP(), apiReq.getAction(), new Date(), "");
+        if (articleService.updateSupportAndDisagreeState(apiReq.getArticleId(), agree, disagree)) {
+            resp.success().setData(true);
+        }
         return resp;
     }
 }
